move existing implementation to utils.c
